Stash compressor state when container closes instead of in localOpMetadata (#17535)
## Description
Stashed ops have been a kludge for IdAllocation ops since the
integration of the IdCompressor into the container runtime. The previous
scheme was to serialize the compressor into the localOpMetadata whenever
an IdAllocation op was generated and to serialize that localOpMetadata
into the ops when `closeAndGetPendingLocalState` was called. On load,
any time a stashed IdCompressor op was seen we would deserialize the
stashed compressor and throw away the compressor.

Rather than do it this way, I've added a compressor entry to
`IPendingRuntimeState` and am serializing the compressor state when this
is generated. Because of the way stashed ops work, there's another
kludge here to skip applying "savedOps" to the compressor as they will
have already been processed at the time of stashing the compressor by
adding a `savedOp` property to the metadata.

This work is the result of trying to integrate using the compressor into
SharedTree, and as a result I discovered a bug that finalizes ranges out
of order if a resubmitted op generates a new id. I've introduced a new
BatchManager and am moving the IdAllocation ops to the front of
processing (similar to attach ops) to solve this problem.

## Breaking Changes
These are technically breaking changes, but since the compressor is
experimental I think these changes should be fine to go into main.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

- Is there a better way to indicate a saved op so that the compressor
can skip them?
- After the next->main merge it looks like the type of metadata has
changed from `any` to `unknown`. While this seems like the right
direction, it seems like `Record<string, unknown>` makes more sense?