GC: Added tests that demonstrate unexpected tombstone behavior and logs (#18092)
Currently, we are seeing some unexpected behavior in the tombstone
feature where tombstone revived are incorrectly logged and in some
cases, object are unexpectedly tombstoned and their unreferenced
timestamp is reset.
Note: The incorrect behavior doesn't lead to incorrect deletion or
tombstoning. It leads to some tombstone revived telemetry and resetting
of tombstone state but these objects will get tombstoned again (i.e.,
they are still unreferenced, its just that the timestamp is reset).
This PR adds tests that validate the below three scenarios. Currently.
the outcome is incorrect:
1. A tombstoned data store (ds1) has handle to another tombstoned data
store (ds2). When ds1 is loaded in an interactive client, a tombstone
revived event is logged for ds2.
2. A tombstoned data store (ds1)'s DDS has handle to another tombstone
data store (ds2). When ds1's DDS is loaded for the first time in
summarizer client, a revived event is logged for ds2 and it gets
untombstoned. ds2 is still unreferenced but its unreferenced timestamp
is reset.
3. A tombstoned data store (ds1) has 2 DDSes. dds1 has handle to dds2.
When dds1 is loaded for the first time in summarizer client, a revived
event is logged for dds2 and ds1 gets untombstoned. ds1 is still
unreferenced but its unreferenced timestamp is reset.

Repros the bug in
[AB#5976](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5976)