GC: Add GC configs to inactive / tombstone telemetry (#18135)
There are certain GC configs and flags that are used to analyze and
filter telemetry such as GC version, tombstone configs / flags,
inactive, sweep and session expiry timeouts. These are not logged along
with inactive object / tombstone telemetry which makes it a little
complex to get these values from other logs.
This PR adds all GC configs to the inactive / tombstone logs so that
analyzing and filtering this data becomes simpler. It also adds
tombstone and sweep flags to the tombstone logs.


[AB#6136](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/6136)