GC: Add a GC option to control throwing on loading a tombstoned object (#18185)
Currently, throwing on loading a tombstoned object is disabled by
default. It can only be enabled via config flags. Since config provider
is passed to the loader, it means this behavior can be controlled at the
loader layer not at the container layer.
This is not always desirable for all customers and some of them may want
to control this behavior at the container level. This will also let them
progagate changes to this behavior quickly since container changes are
propagated faster.