build: Run readme generation in build (#18198)
The scripts and tools we use to generate readme contents in the repo has
never been run as part of the main build. This PR adds a step to the
build, `build:readme` which calls markdown-magic to regen the readmes
across the repo, excluding the docs folder.

There are similar scripts in the docs package. Those don't need to be
modified; they're just a little unnecessary now.

I also updated the repo-wide policy CI workflow (the one that runs on
every PR) to no longer run the root prettier script (it's now part of
the build) and runs build:readme instead. This will catch any changes
outside the client release group, since md-magic runs repo-wide.