build(client): Use new node16 tsconfigs in azure, client, and DDS projects (#18152)
This PR is the first in several that will update the CommonJS builds
within the repo to use `module: node16`, which implies
`moduleResolution: node16`.

There are new configs in the build-common package, and these configs are
used as follows:

```json
"extends": [
	"../../../../../common/build/build-common/tsconfig.base.json",
	"../../../../../common/build/build-common/tsconfig.cjs.json",
],
```

These configs are designed to be used together using [TypeScript's
support for extending multiple config
files](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-0.html#supporting-multiple-configuration-files-in-extends).

- tsconfig.base.json - This base config contains defaults that all
packages within the repo should use as a baseline.
- tsconfig.cjs.json - This config sets `module: Node16` and is intended
for CommonJS builds. This config is intended to
  be layered on top of the base config.
- tsconfig.esm.json - This config sets `module: ESNext` and
`moduleResolution: Node10` and is intended for ESM builds. This config
is intended to be layered on top of the base config. This config uses
Node10 module resolution because Node16 cannot be used to build both CJS
and ESM from a common source.
- tsconfig.test.json - This config disables some settings that we don't
want to use in test code, like `declaration` and `decarationMap`. It
also enables the `node` types by default.

This PR updates the packages in the packages/common and packages/dds
paths. For each package, the CJS, ESM, and test tsconfig files are
updated to use the new build-common configs, which means that CJS builds
now use module: node16 which implies moduleResolution: node16. ESNext
builds continue to use Node10 module resolution because Node16 cannot be
used to build both CJS and ESM from a common source.