feat(check:policy): Make policy handlers async (#17931)
Policy handlers are now async, which enables them to more easily use
async libraries to perform checks. Note that resolvers and "final"
handlers are still synchronous.

I also updated the names of some functions and variables to make it
clearer what they're doing. In particular the "runPolicyCheck" function
was very wrong - it only executed the final handlers so it's a lot
clearer post-rename.