Enable dts trimming for some packages (#18176)
This PR updates a handful of packages to use dts trimming. As far as we
know, the rollups don't differ between CJS and ESNext builds, so we just
copy the rollups from the CJS build into the ESNext build output.

If we find this causes problems, then the alternative is similar to what
we do with tsconfigs - we create a new config file,
api-extractor.esnext.json, and then we add a task to package.json to run
api-extractor using that new config file. This is the most "correct" way
to do this since it runs api-extractor both times, using the CJS and
ESNext source as appropriate. However, that method is slow because it
calls api-extractor twice, so we're going to use the copy method until
we find we can't.

This PR does not change to use a trimmed type file, like public. We can
make those changes separately. This PR just enables building the trimmed
files.


[AB#6146](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/6146)