clientSequenceNumber cleanup (#17959)
For `PendingStateManager` change, a cleanup item was never opened to
remove the property from `IPendingMessage`
https://github.com/microsoft/FluidFramework/pull/14434.

For `OpGroupingManager`, the `clientSequenceNumber` would currently
represent 2 separate concepts: index in batch, and actual
clientSequenceNumber. We should only present one of these concepts at a
time depending on if grouped batching is enabled or not.