server: Add Nexus to server/docker-compose (#18228)
## Description

The new Nexus service was not added to the server/docker-compose*.yml
files, so running R11s that way does not work with AzureClient because
it can't find Nexus. This isn't terrible, since I think I'm one of very
few people who use docker-compose.dev.yml, but would be nice to have it
fixed.