refactor(infra): Rework getRequestedRange to return exact version used (#18201)
## Description

This PR reworks the existing `getRequestedRange` function into
`getRequestedVersion`. The main change is that it returns the latest
exact version requested instead of a semver range. This will be helpful
when debugging compat testing since it will be more clear which version
was used in testing.

Logic pulled out form
https://github.com/microsoft/FluidFramework/pull/17158 /
https://github.com/microsoft/FluidFramework/pull/18087

## Misc 

[AB#5929](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5929)