tree2: Cleanup typed/simple tree entry point (#18255)
## Description

Cleanup typed/simple tree entry point to actually use a wrapper object
for the view and expose what's needed for undo+redo and disposal.

This should allow apps to use this simpler entry point.

It also makes it possible to refer to TreeView types without needing
access to the schema types, since its now parametric over the Proxy type
instead.

This is work toward having the "simple" public layer be more of a
wrapper and less of an API subset, which better aligns with the
differing data model and concepts between the public API and the
implementation logic.

The changes to the inventoryList sample show the main intended impact of
these changes.

## Breaking Changes

Code explicitly providing type parameters to TypedTree view will need to
provide the proxy API instead. Users of the internal types and
"schematize" need to call schemtizeInternal now. The typed tree types
have been renamed.