Cluster capacity request (#18262)
## Description

This change annotates the compressor creation ranges with the desired
cluster size. This allows for future "smart" request policies without
needed to make a format change, as the policy can be run by one client
without sequencing.

## Breaking Changes

This breaks the IdCompressor persisted format, but no code yet uses it
so this is acceptable.