Use migration shim in version-migration example (#18215)
In the video you can see two connected clients, the top one and the
bottom one. We click migrate on the button and it goes from legacy
shared tree to new shared tree. Both locally and remotely


https://github.com/microsoft/FluidFramework/assets/90650728/a79db7a8-1526-48ba-8263-4c950096affd

This adds the usage of both the `MigrationShim` and the `SharedTreeShim`
to the `examples/version-migration/tree-shim` example.

## Notable changes:
- adds `MigrationShim`, `SharedTreeShim`
- Uses one `IFluidHandle` entry instead of two, see `treeKey`
- `isMigrated` is based on the type of the `shim`
- reacting to the migration is setup in the `hasInitialized` function of
the `InventoryList` object.

## Future improvements
- A more in depth tree migration
- A way to guarantee v1 ops are sent after the barrier op