build(azure): Use tsc-multi to dual-emit CJS and ESM (#18273)
Updates the azure packages to use
[tsc-multi](https://github.com/tommy351/tsc-multi) to build ESM and CJS
with node16 module resolution.

Since this is the first time using tsc-multi, I added some shared
configs to build-common that will be used by all packages as we roll out
these changes. Except for the test config, which exists in every project
because of bugs/odd behavior in tsc-multi. We nest our test tsconfigs
within the src folder, and no combination of tsc-multi flags worked
unless the config file is local to the package. This seems like a fine
workaround for now, but long term we'll probably want to do something
else.

I'm deliberately holding off on writing a bunch of docs until we've
verified that this approach will work across more packages.