Revert "build: Use link: protocol instead of file: (#18270)" (#18286)
This reverts commit be5d92d2f02dc0b7e093e15b3eb1d5b52e563d56.

Based on the documentation
[here](https://pnpm.io/cli/link#whats-the-difference-between-pnpm-link-and-using-the-file-protocol),
`file:` actually seems the more appropriate option over `link:`. There
are some issues with `file:` that still need to be investigated (namely
the claim that source-code modifications are picked up automatically,
which does not always seem to be the case), but there are more immediate
issues with `link:` that we wish to avoid.