fix(fluid-build): Include source files in tsc-multi done file (#18292)
The done file that fluid-build was using for tsc-multi task caching did
not include the source or config files, so source changes did not cause
a rebuild like they should. This PR updates the task to include those
source files in the done file calculation, which means that fluid-build
will correctly trigger the tsc-multi tasks when source files change.