[main > release/v2int/7.0]: Make improvements in calculating retry after time for calls to storage/relayService (#18315)
## Description

Make improvements in calculating retry after time for calls to
storage/relayService. The logic at service side is not very sound to
calculate this value. So at least in short term, improving this logic at
client side to be intelligent in deciding this value.
Loop faced a bug related to this, so we need to provide a short-term
viable sol.