tree2: Setup Simple layer (#18329)
## Description

Move simple-tree out of feature libraries.
Also rename SharedTreeObjectFactory to TreeObjectFactory.

See update to ReadMe for changes to the dependency structure.

The idea here is that simple-tree should not be depended on by
shared-tree-core (or anything it uses), but instead is a wrapper around
the system exposing its simpler types and simpler data model.

This change makes it possible/logical (and cycle free) for a simple-tree
to contain the leaf domain, and a specialized schema builder. This opens
up a lot more options for how the schema builder can be designed and
exposed.

## Breaking Changes

Rename SharedTreeObjectFactory to TreeObjectFactory.