Prevent SharedTree from sending ops if the change in the op is invalid (#18377)
## Description

Invalid changes, though not expected, can and will arise due to bugs in
editing or forest mutation code. In the event that a change made by a
client crashes the client, it should not submit the op containing that
change (and making all other clients crash too after that point in the
document history). Currently changes are assumed to be valid and are
always submitted. We do not currently have a general mechanism for
validating a change, therefore, this PR changes the application of the
change on the local client to happen before the submission of the op.
That way, if the local client crashes because the change is bad, op
submission will not happen.