Add Strong Typing to Interval Collection Ops (#18389)
Interval collection ops didn't have strong typing all the way through
due to legacy coupling to value type infrastructure. The coupling is
still there, but the value type infrastructure isn't used generically,
it is only used for interval collections. this change changes the typing
to reflect that, and it makes the interval collection code easier to
follow.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>