Deprecate FluidStatic Classes (#18402)
Deprecate FluidStatic Classes

This change deprecates a number of unnecessarily exposed FluidStatic
classes, and replaces them with creation functions. This helps us keep
implementations decoupled from usage which is easier to maintain and
extend. It has very minimal impact on the public surface area of
downstream packages. The deprecated classes are as follows:

-   AzureAudience
-   TinyliciousAudience
-   DOProviderContainerRuntimeFactory
-   FluidContainer
-   ServiceAudience

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>
Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>
Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>