(odsp-client): Interface cleanup and pass itemId to getContainer (#18387)
Couple of feedback which are included in this PR are as follows:

1. Pass itemId to load the container instead of absolute url
2. Remove other unused interfaces

Remaining changes to the interface - 

We discussed returning the `fileName` to the app developer. The
resolvedUrl returns fileName as an empty string. fileName is not part of
the current interface, but there will be a follow-up for this change.