feat(build-common): Add `ae-undocumented` rule to base API-Extractor config (#18404)
The `ae-undocumented` rule enforces that items bundled in the package
exports rollup contain TSDoc comment content.

For now, the rule is enabled to log violations to the generated API
report documents, rather than failing the build. We will consider
promoting this to a build failure in the future, and packages can opt
into build failure policy as desired in the interim.