Stop sending ops while closing and getting pending state (#18188)
[AB#5127](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5127)
We're disconnecting the container while getting pending ops, to be able
to still send ops to the pending state manager but not submit them,
given that we are about to close. That approach was not the most
appropriate one because in any given moment, the container can reconnect
and start sending ops again.

This change removes disconnectInternal from closeAndGetPendingLocalState
and acomplishes the same by just using canSendOps function in
containerRuntime, setting it to false if we are getting pending state
and about to close the container.