tree2: Simplify schema aware modes (#18424)
## Description

SchemaAware and contextually typed data are now only used at the
flex-tree layer.
As this layer now has the fallback option of cursors, its ok of some of
the edge cases that the low level schema can express don't have nice
strongly typed literals (the edge case in question is ambiguous leaf
schema, like union of node key or string). Thus support for the flexible
API mode has been removed, and only "simple" is kept. As this is the
only mode, the mode selection enum was also removed.

Additionally, since its simpler now, and a test stopped compiling due to
it (it's unclear why it worked before), support for map nodes was added.

## Breaking Changes

Users of explicitly typed leaf nodes will have to just provide the leaf
value or use a cursor.