build: Update API-Extractor configs to extend base config by file path, rather than via package dependency (#18428)
We are attempting to move away from package dependencies as a pattern
for implementing build tooling. This PR updates our API-Extractor
configs to extend our shared base config via relative file path instead
of via the `build-common` package dependency.

Doing so pulls in the most recent changes from that config, which have
not yet been published. In particular, it pulls in this change: #18404
which adds a requirement that all exported package members be
documented. For now, I have disabled the rule in any packages that
reported errors ~~opting to log those errors to the generated report
rather than failing the build~~. We will address fixing those violations
in the future. ~~For now, devs will at least be alerted when they add
new undocumented items via new warnings in the report.~~

UPDATE: it turns out that api-extractor logs line numbers for all
missing docs violations. This is extraordinarily susceptible to code
churn. To avoid creating unnecessary noise in PRs, I've removed the
report logging in this PR.