refactor(merge-tree sequence): remove combining op and related codepaths (#18096)
Removes ICombiningOp and code gated on different values of it. This code
has no code coverage and is generally unused by partners.

Changeset+deprecation PR coming after CI passes. 

I'm considering modifying some of the APIs to accept `undefined` in the
place of the combining op to reduce code churn, but based on a quick
glance of partner code, this likely won't be necessary or useful.