PropertyDDS: Summary generation Bug Fix (#18072)
## Description
This PR fixes a bug that resulted in the prune algorithm in PropertyDDS
to prune remoteChanges but keep the associated unrebasedChanges when
generating a summary. This could later result in rebase errors for
clients that would join using this summary as a starting point.

The problem would occur, when a summarizer has several remote changes in
its history, but the MSN points to a state before the first of these
changes. In that case, none of those can yet be removed, but the old
implementation would incorrectly handle this case and remove the remote
change, but keep the unrebasedChanges.

## Breaking Changes

This change only affects the summarizer and thus will not cause any
incompatibility between clients with different versions.