Add capabiity to create channel first and then supply it to data store runtime (#18418)
## Description

There have been requirements to have ability to create channel with
custom properties related to a particular DDS and then attach that to
the data store runtime. This comes in handy in cases let's say when you
need to supply some policy to a DDS, or any prop and also have types
checking instead of any. Some of these requirements are in SharedTree or
SharedMatrix etc.
So, this PR achieves that by overloading the` createChannel` API in
dataStore Runtime class which can now directly take in a channel instead
of creating inside it.

This PR is based on this PR by @anthony-murphy :
https://github.com/microsoft/FluidFramework/pull/17783/files
Also added unit tests for the same.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>
Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>