(odsp-client): Replicate ITokenProvider in odsp-driver (#18474)
This PR replicates the ITokenProvider implementation in odsp-driver.
Currently, ITokenProvider is part of routerlicious-driver. Importing
ITokenProvider from routerlicious-driver for odsp-client seems
inappropriate. Therefore, this PR replicates the implementation,
altering the return promise signature from ITokenResponse to
TokenResponse.

In the process of making this modification, the azure-client dependency
is also eliminated.


[AB#6338](https://dev.azure.com/fluidframework/internal/_workitems/edit/6338)


[AB#6275](https://dev.azure.com/fluidframework/internal/_workitems/edit/6275)