tree2: Class based schema (#18350)
## Description

To support recursive schema using classes for the schema is useful to
work around https://github.com/microsoft/TypeScript/issues/55832 .

If we are going to use classes to enable that workaround, and also
create a separate simple-tree specific schema system, there are some
other interesting approaches which can be taken.

If we don't support "Any" as a type allowed in fields, then all schema
are reachable from the root.
This means we can skip the schema library concept, and just pass in the
root schema in the configuration.
This has the side-effect of making the actual schema objects available
programmatically be the actual class objects declared in the schema file
(instead of the classes which those extend, which is what a library
would have captured).
This same reachability from the root also opens the door to supporting
contextual view schema where the same type in different locations in the
tree could have different schema subclasses, and its parent would
unambiguously determine which to use for it.

Thus all together this approach makes it possible to make the tree nodes
instances of classes which the user declared.
This has several benefits:

- Users don't have to use "typeof" or invoke any type meta-functions to
get the node types they want to pass around: just use the class/schema
name as the type.
- Intelisense is much cleaner when referring to types defined in schema:
It just uses the class name. or "typeof ClassName". (These
simplifications are what resolve the d.ts issue noted above with
recursive types)
- normal JS/TS type narrowing with `instanceof` can be used with schema
defined types.
- Its possible to add view/session local state to instances as
properties, as well as adding methods by just putting them in the class
like any other class.
- Since the schema/class defines the API for the node, the
implementation of the tree API implicitly gets defined in a way that is
trivially extensible for adding new node kinds.

TODO:

- [X]  support recursive times.
- [x]  implement runtime behavior for object fields.
- [X]  implement runtime behavior for list indexing.
- [x]  structurally typed maps and lists.
- [x] runtime constructing of hydrated nodes (ensure flex tree is type
erased).
-  runtime constructing and later hydrating of unhydrated nodes.
   - [x] objects
   - [ ] map
   - [ ] list
- [x]  Generation of view schema from simple schema.
- [x] Implement full simple-tree API surface
- [ ] Evaluate impact on simple-tree API (for example exposing of
prototypes impacting deep equals).
- [x] Basic end to end testing
- [ ] Full unit testing.

## Breaking Changes

The `schematize` API for simple-trees not using the new class based APIs
has been renamed to `schematizeOld`:
either rename the call site or migrate to the new class based schema.

Some types which would have collided have their old simple-tree or
flex-tree non class based version's renamed:

```typeScript
TreeNodeSchema as FlexTreeNodeSchema,
TreeListNode as TreeListNodeOld,
Tree as TreeOld,
TreeApi as TreeApiOld,
TreeView as TreeViewOld,
```

The previous version of all these names are now used for class based
APIs.