Tree: Fix bug with determining relative order of concurrently inserted cells (#18422)
## Description

This PR fixes a bug where, when rebasing a branch with multiple changes,
we were in some cases unable to determine the relative order a cell
created on the source branch and a cell created on the target branch.
This is now resolved by using the relative order of the cells'
revisions. The cell with the later revision should come before the cell
with the earlier revision. This information must now be provided through
a RevisionMetadataSource to the top level call to the ChangeRebaser.

---------

Co-authored-by: yann-achard-MS <97201204+yann-achard-MS@users.noreply.github.com>