GC: Updated GC stats to include deleted and life time stats (#18506)
Update the "GarbageCollection_end" event to include the following stats:
- deletedNodeCount - The total number of nodes deleted in a container.
- deletedDataStoreCount - The total number of data stores deleted in a
container.
- deletedAttachmentBlobCount - The total number of attachment blobs
deleted in a container.
- lifetimeNodeCount - The total number of nodes in the life time of a
container.
- lifetimeDataStoreCount - The total number of data stores in the life
time of a container.
- lifetimeAttachmentBlobCount - The total number of attachment blobs in
the life time of a container.
- sweep - Whether sweep ran or not. This will help distinguish the
delete stats between "actually deleted" and "potentially deleted".

This will help us analyze how many objects GC has deleted throughout the
lifetime of the document and we can include that in the GC dashboards.

Currently, the stats for deleted objects is logged via
"GC_SweepReadyObjects_Delete" event. However, it has a bug where its not
always logged and it only include the list of deleted ids. Removed this
log in this PR.

AB#6362