return deep copy of client from loader options (#18246)
Alternative to previous closed PR:
https://github.com/microsoft/FluidFramework/pull/17479

This PR is intended to address a bug that occurs during
disconnect/reconnect scenarios where a client gets unintentionally
mutated with a "non-interactive" attribute during the reconnect process.
By returning a deep clone of the client, these unexpected mutations can
be prevented.

Previously, a similar fix was implemented using lodash deepclone,
however this resulted in an undesired increase in bundle size. Since
then, we have moved beyond the minimum required TypeScript version bump
to >4.7 and node version >17, which allows us to use the built-in
structuredClone method.

Note that in the original PR
https://github.com/microsoft/FluidFramework/pull/17479, the options
variable was deep-cloned instead of the client. However, this causes
some e2e tests such as attribution to fail since attribution passes in
complex function objects into options which is not supported by the
structuredClone function used.

Testing the changes on the external controller example, this method
addresses the bug as well since the mutation seems to mainly be
occurring on the client attributes.

Here is the browser compatibility as well as inputs which are not
supported by structuredClone from the documentation
https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm
<img width="822" alt="Screenshot 2023-11-16 at 2 03 01 PM"
src="https://github.com/microsoft/FluidFramework/assets/112977307/1b10d071-0d97-4917-94a5-0837ceff290c">
<img width="752" alt="Screenshot 2023-11-16 at 2 07 50 PM"
src="https://github.com/microsoft/FluidFramework/assets/112977307/b0272c20-c8b2-4763-a22e-f8de36dfd8f3">

In addition to deep cloning the client, the original `get client()` was
replaced with a `readonly client` property which is set in the
constructor using the previous implementation of the getter as a helper
function. This will prevent structuredClone from creating copies
everytime get client is called.