When inserting into a SharedTree, only hydrate proxies, not plain objects (#18560)
## Description

There is currently a bug when inserting "plain" JavaScript objects (as
opposed to "raw" proxies created with content factories/constructors).
The plain object will undergo "hydration", which is something only
applicable to proxies. This causes the plain object to be associated
with the contents of the tree, and thus subsequent reads of the tree
will produce that content object rather than creating a new proxy. This
PR adjusts the logic to only hydrate objects which are indeed already
thirsty proxies.