build(client): Add `check:release-tags` to enable enforcement of release tag compatibility across package boundaries within the repo (#18536)
## Overview
Actual enforcement of release tag compatibility will be enabled in the
future, once we've gone through and tagged the entire repo. This PR
enables the linter step, but for now it doesn't validate anything beyond
what our api-extractor build stage already validates.

AB#6320

## Background
A separate linter step is currently required because we have made the
explicit choice not to require bundling all required members of local
dependencies in packages that surface types which (transitively)
reference them. But we still need to ensure that transitively referenced
API members are available at the right release scope for our customers
to use.

### Example

Consider the following package relationship:

Package `Foo`

```typescript
/**
 * @internal
 */
export interface IFoo {...}
```

Package `Bar` (depends on `Foo`)

```typescript
import { IFoo } from "Foo";

/**
 * @public
 */
export function bar(foo: IFoo) {...}
```

We have made the choice that we are okay requiring users of the `Bar`
package to also import the `Foo` package in order to get access to the
`IFoo` type consumed by the `bar` function. While this can be convenient
to the end-user, our framework has many layers with many cross-package
references, which would cause our APIs to quickly explode in scope.

***BUT*** We need to ensure that `IFoo` is actually available for those
consumers to find, otherwise we have introduced a function that is
impossible to use in a type-safe way.

In the above example, you'll note that the `IFoo` interface is marked as
`@internal`, but we are exporting the `bar` function (which accepts an
`IFoo` parameter) as public. In this scenario, a user of `Bar`'s public
API would be unable to find the type definition for `IFoo` in the `Foo`
package, as we have marked it as internal to the framework, and so it
will be trimmed out of our public API in the published package.

By default, API-Extractor does not do any cross-package analysis to
validate scenarios like this. This makes sense - who knows how our
myriad third party dependencies have tagged their APIs?

*Except* API-Extractor does expose a feature for bundling other
packages' API members. This feature is aimed at making it possible to
automatically bundle dependent APIs directly in a given package. E.g.
automatically bundling `IFoo` into the `Bar` package so that users don't
have to take a direct dependency on `Foo`. This feature performs all of
the normal API validation run within a package across all of the
specified dependencies, so it will also catch release tag
incompatibilities across package boundaries.

In lue of any explicit tooling on the API-Extractor side for _only_
validating cross-package release tag compatibility, we have enabled a
separate "lint" run of API-Extractor that turns off all of the build
aspects of the normal run (metadata and api report generation,
type-rollup generation, etc.) and strictly runs validation of select
rules we want to enforce across package boundaries.

## API-Extractor Patch

API-Extractor is currently missing a couple of changes that we need to
implement this cleanly and correctly. These changes have been captured
via a pnpm patch to API-Extractor in the repo. Details can be found in
[patches/README.md](https://github.com/microsoft/FluidFramework/blob/afa436fd763524c2fa9084a48918daad868293cf/patches/README.md#microsoftapi-extractor).

## For reviewers

Note that while there are many file changes here, the majority of them
are one of the following:
1. Adding the `api-extractor-lint.json` file to the package
2. Adding the relevant "check:release-tags" scripts to the
`package.json`

Actual changes of interest are in the following files:
1. `fluidBuild.config.cjs`
2. `package.json`
3. `common/build/build-common/api-extractor-lint.json`
4. `patches/*`

## Current Limitations

There are a couple of limitations that we may or may want to address now
(given our time constraints). Calling them out here for transparency /
discussion.

1. Each package is currently opting into the cross-package validation by
adding the `check:release-tags`. There is currently nothing enforcing
that our library packages contain this script, so issues could fall
through the cracks if a library package was added without the script, or
if I missed one here.
* In theory we could add a policy-check handler to enforce that
`check:release-tags` is present if the package.json does not include
`"private": true`.
2. The enforcement is currently only enabled for the client release
group (`azure`, `experimental`, and `packages`).
* It would be easy enough to add to the others if we want, but I didn't
think it was as urgent as tackling `client`.