(odsp-client): Setup e2e test cases  (#18550)
This PR establishes end-to-end test cases that mirror those in
azure-client. While the test cases remain identical, the distinction
lies in the client configuration and token provider used.


![image](https://github.com/microsoft/FluidFramework/assets/48232592/f552cf6c-161f-4f00-a69e-9bc230e435a8)


List of remaining/followups work:
1. Currently, the audience test has asserts commented, and the size of
members always remains at 1, as a single M365 account is utilized for
login. When a new test tenant is generated, these tests can be modified
accordingly.
[AB#6425](https://dev.azure.com/fluidframework/internal/_workitems/edit/6425)
2. Fix pending test (DONE)


[AB#6390](https://dev.azure.com/fluidframework/internal/_workitems/edit/6390)